requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-ollama
langgraph
sqlalchemy
pydantic
python-dotenv
python-dateutil
alembic
redis
langchain-openai
langgraph-checkpoint-sqlite